Refactor

+++ src/main/java/kr/hhplus/be/server/application/order/OrderResult.java
@@ -1,8 +1,9 @@
 package kr.hhplus.be.server.application.order;
 
-import kr.hhplus.be.server.domain.order.OrderInfo;
 import kr.hhplus.be.server.domain.payment.PaymentInfo;
 import lombok.Builder;
+
+import java.math.BigDecimal;
 
 public class OrderResult {
 
@@ -16,7 +17,7 @@ public class OrderResult {
     public record Payment(
             Long orderId,
             Long userId,
-            int paymentAmount
+            BigDecimal paymentAmount
     ) {
         @Builder
         public Payment {}
