+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/PointResponse.java
@@ -3,6 +3,8 @@ package kr.hhplus.be.server.interfaces.api.point;
 import kr.hhplus.be.server.domain.point.PointInfo;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 public class PointResponse {
     public static Point of(PointInfo.Point point) {
         return Point.builder()
@@ -13,7 +15,7 @@ public class PointResponse {
 
     public record Point(
             Long userId,
-            int point
+            BigDecimal point
     ) {
         @Builder
         public Point {}
